backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis Cache
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# Email
aiofiles==23.2.1
jinja2==3.1.2

# Monitoring and Logging
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
python-json-logger==2.0.7

# Task Queue
celery==5.3.4
celery[redis]==5.3.4

# Environment and Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Development Tools
python-multipart==0.0.6
email-validator==2.1.0

# WebSocket Support
websockets==12.0

# Data Validation
email-validator==2.1.0

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Rate Limiting
slowapi==0.1.9

# Utilities
uuid==1.30
python-dateutil==2.8.2
pytz==2023.3